// src/modules/resource/types/AlternanceAndFormation.ts
export interface AlternanceAndFormation {
  diplomaLevel: string;
  period?: any;
  capacity?: any;
  ville: string;
  departement: string;
  zipcode: string;
  emailContact?: any;
  romeDomain: string;
  entreprise: string;
  _id: number;
  reference: string;
  _title: string;
  id_Category: number;
  _activation: boolean;
  modification?: any;
  _user: number;
  _views: number;
  shared: number;
  _creationDate: string;
  _url?: any;
  _comments?: any;
}