// src/modules/resource/types/Job.ts
export interface Job {
  description: string;
  experience: string;
  ville: string;
  salaire: string;
  zipcode: string | null;
  typeContrat: string;
  codeNaf: string;
  _id: number;
  reference: string;
  _title: string;
  id_Category: number;
  _activation: boolean;
  modification: string | null;
  _user: number;
  _views: number;
  shared: number;
  _creationDate: string;
  _url: string;
  _comments: string | null;
}